+++ src/com/clouway/jdbc/customer/history/persistence/CustomerRecord.java
@@ -6,14 +6,14 @@ import java.sql.Timestamp;
  * @author Krasimir Raikov(raikov.krasimir@gmail.com)
  */
 public class CustomerRecord {
-    public final int recordId;
+    public final long recordId;
     public final Timestamp changeDate;
-    public final int customerId;
+    public final long customerId;
     public final String name;
     public final String lastName;
     public final String egn;
 
-    public CustomerRecord(int recordId, Timestamp changeDate, int customerId, String name, String lastName, String egn) {
+    public CustomerRecord(long recordId, Timestamp changeDate, long customerId, String name, String lastName, String egn) {
         this.recordId = recordId;
         this.changeDate = changeDate;
         this.customerId = customerId;
